ai-code-generator/requirements.txt
transformers==4.30.0
torch>=2.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pytest>=7.0.0
python-multipart>=0.0.5
spacy
python-magic
pygments
esprima
torch==1.10.0
transformers==4.15.0
datasets==1.17.0
tqdm==4.62.3
pandas==1.3.4
numpy==1.21.4
scikit-learn==1.0.1
matplotlib==3.5.0
seaborn==0.11.2
google-generativeai==0.1.0
openai==0.27.0
tenacity==8.0.1
mlflow==1.22.0
wandb==0.12.9
pytest==6.2.5
flake8==4.0.1
black==21.12b0
isort==5.10.1